% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_column_names}
\alias{all_column_names}
\title{Column names}
\format{
An object of class \code{list} of length 3.
}
\usage{
all_column_names
}
\description{
List of column names used in the code, from the datalist and exportfiles
in all four languages (de, fr, it, en)
}
\keyword{datasets}
