% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{build_custom_mapping}
\alias{build_custom_mapping}
\title{Build column name mappings}
\usage{
build_custom_mapping(data, language = "de", prompt_mapping = TRUE)
}
\arguments{
\item{data}{the custom dataframe for which the user wants to build a custom
mapping}

\item{language}{a character string representing the language in which the
columns will be displayed during the mapping prompt (\code{"de"} or
\code{"fr"} or \code{"it"} or \code{"en"})}

\item{prompt_mapping}{a boolean indicating whether the function prompts the
user for the exact mapping of his dataframe or whether the columns are
mapped automatically by order}
}
\value{
TODO: Oh boy... a vector of column names indexed by column names... make
that clear
}
\description{
\code{build_custom_mapping} creates a vector of column name mappings for the
user to read her or his custom dataframe
}
\details{
TODO: A longer description of what is being done.
}
