% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{build_errors}
\alias{build_errors}
\title{Builds a dataframe of errors}
\usage{
build_errors(rows, pers_id, vals, column, description, importance)
}
\arguments{
\item{rows}{a vector of numbers representing the rows which contain an error}

\item{pers_id}{a vector of strings of the personal ID which contain an error}

\item{vals}{a vector of the erroneous values}

\item{column}{the name of the column containing the error}

\item{description}{the description of the error occurring}

\item{importance}{the importance of the error occurring}
}
\value{
a dataframe of errors with the columns \code{column},
\code{description}, \code{importance}
}
\description{
\code{build_errors} builds a dataframe of errors as used by the function
\code{check_data}.
}
\keyword{internal}
