% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{check_data}
\alias{check_data}
\title{Check a dataframe}
\usage{
check_data(data)
}
\arguments{
\item{data}{data.frame to be checked}
}
\value{
a data.frame with information concerning each incorrect data point
in the \code{data} data.frame
}
\description{
\code{check_data} checks a dataframe (as produced by
\code{\link{read_data}}).
}
\details{
This function checks a dataframe (as produced by \code{read_data}
for correctness and consistency)
}
\keyword{internal}
