% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transformation.R
\name{compute_age}
\alias{compute_age}
\title{Compute age values}
\usage{
compute_age(x, age_spec = NULL, reference_year = NULL)
}
\arguments{
\item{x}{a string or number vector to be transformed}

\item{age_spec}{a string indicating the age specification, can be one of
\code{NULL}, \code{"age"}, \code{"birthyear"}, or \code{"date_of_birth"}. If
this parameter is set to \code{NULL}, the function automatically tries to
infers the specification}

\item{reference_year}{a number indicating the reference year in order to
compute the age from a birth year or birth date. If \code{age_spec} is
\code{"age"}, this parameter can be ignored.}
}
\value{
a numeric vector of ages
}
\description{
Computes the age given a birth year or a birth date
}
\keyword{internal}
