% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalists.R
\name{download_datalist}
\alias{download_datalist}
\title{Download official Excel datalists}
\usage{
download_datalist(file, language = "de")
}
\arguments{
\item{file}{a character string representing the file path to which the
downloaded datalist will be saved.}

\item{language}{a character string representing the language of the datalist
to be download (\code{"de"} or \code{"fr"} or \code{"it"} or \code{"en"}).}
}
\value{
None
}
\description{
Downloads an empty version of the latest official excel datalist in the
specified language to the given \code{path}.
}
