% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{feb29_between}
\alias{feb29_between}
\title{Check if the interval between two dates contains February 29}
\usage{
feb29_between(date1, date2)
}
\arguments{
\item{date1}{the first date by chronological order}

\item{date2}{the second date by chronological order}
}
\value{
a boolean
}
\description{
Check if the interval between two dates contains February 29
}
\keyword{internal}
