% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{map_data}
\alias{map_data}
\title{Reads a dataframe and maps it columns}
\usage{
map_data(data, data_origin = "datalist", custom_map = NULL)
}
\arguments{
\item{data}{data.frame to be transformed}

\item{data_origin}{a string indicating the origin of the data, can be one of
\code{datalist}, \code{data_export}, or \code{custom}}

\item{custom_map}{a character vector as produced by
\code{build_custom_mapping}}
}
\value{
a data.frame with the column names as used in the code
}
\description{
\code{map_data} reads a dataframe and maps it columns to match the ones
used in the salary analysis.
}
\details{
Renames the columns a data.frame to the column names used in the standard
analysis model. The mapping is based either on the origin of the data (i.e.
official datalist or data export from the webtool) or using a custom map
as produced by \code{build_custom_mapping}
}
\keyword{internal}
