% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalists.R
\name{read_official_excel}
\alias{read_official_excel}
\title{Read official datalist or data_export Excel file}
\usage{
read_official_excel(path)
}
\arguments{
\item{path}{a character string indicating the path of the Excel file to be
read}
}
\value{
a dataframe with the contents of the datalist or data_export
}
\description{
Reads an official datalist or data_export file into a dataframe object.
}
\keyword{internal}
