% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{Run the Salary Analysis}
\usage{
run_analysis(
  data,
  reference_year,
  female_spec = "F",
  male_spec = "M",
  age_spec = NULL,
  entry_date_spec = NULL,
  ignore_plausibility_check = FALSE,
  prompt_data_cleanup = TRUE
)
}
\arguments{
\item{asdf}{TODO: specify params}
}
\value{
object of type \code{run_analysis} with the following elements
\itemize{
   \item{\code{original_data}}{The original data passed by the user in the
   \code{data} parameter}
   \item{\code{clean_data}}{The cleaned up data which was used for the
   analysis}
   \item{\code{results}}{The result of the standard analysis model}

}
TODO: S3 object, what type is it?
}
\description{
TODO: Description
}
\examples{
TODO: add examples

}
