% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{standard_analysis}
\alias{standard_analysis}
\title{Run a Salary Analysis}
\usage{
standard_analysis(
  data,
  reference_year,
  female_spec = "F",
  male_spec = "M",
  age_spec = NULL,
  entry_date_spec = NULL,
  ignore_plausibility_check = FALSE,
  prompt_data_cleanup = FALSE
)
}
\arguments{
\item{TODO:}{specify params}
}
\value{
object of type \code{standard_analysis_model} with the following
elements
\itemize{
   \item{\code{original_data}: }{The original data passed by the user in the
   \code{data} parameter}
   \item{\code{clean_data}: }{The cleaned up data which was used for the
   analysis}
   \item{\code{results}: }{The result of the standard analysis model}

}
}
\description{
Runs a salary analysis according to the Swiss standard analysis model
}
\examples{
TODO: add examples

}
