% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summary.analysis_model}
\alias{summary.analysis_model}
\title{Summary of the Salary Analysis}
\usage{
\method{summary}{analysis_model}(object, ...)
}
\arguments{
\item{object}{estimated salary analysis object of class
\code{analysis_model}}

\item{...}{further arguments passed to or from other methods}
}
\value{
Nothing
}
\description{
Summary of an estimated salary analysis object of class
\code{analysis_model}
}
\details{
\code{summary.analysis_model} provides a short summary of the wage
analysis according to the Standard Analysis Model. The summary describes the
number of records used for the analysis, the Kennedy estimate of the wage
difference under otherwise equal circumstances and the summary of the linear
regression.
}
\examples{
# Estimate standard analysis model
results <- analysis(data = datalist_imprimerie, reference_month = 1,
   reference_year = 2019, female_spec = 1, male_spec = 2)

# Show summary of the salary analysis
summary(results)

}
