% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summary.standard_analysis_model}
\alias{summary.standard_analysis_model}
\title{Summary of the Salary Analysis}
\usage{
\method{summary}{standard_analysis_model}(object)
}
\arguments{
\item{object}{estimated salary analysis object of class
\code{standard_analysis_model}}
}
\description{
Summary of an estimated salary analysis object of class
\code{standard_analysis_model}
}
\details{
TODO: Description
}
\examples{
# Estimate standard analysis model
result <- standard_analysis(data, 2020)

# Show summary of the salary analysis
summary(result)

}
