% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{within_a_year}
\alias{within_a_year}
\title{Check if the interval between two dates is less than a year}
\usage{
within_a_year(date1, date2)
}
\arguments{
\item{date1}{the first date by chronological order}

\item{date2}{the second date by chronological order}
}
\value{
a boolean
}
\description{
Check if the interval between two dates is less than a year
}
\keyword{internal}
