% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{yearfrac}
\alias{yearfrac}
\title{Time difference between two dates in fractional year terms}
\usage{
yearfrac(date1, date2)
}
\arguments{
\item{date1}{the first date}

\item{date2}{the second date}
}
\value{
fractional years between \code{date1} and \code{date2}
}
\description{
Computes the time difference between \code{date1} and \code{date2} in
fractional year terms. This is equivalent to the YEARFRAC() method from
Excel, with the parameter "Actual/Actual"
}
\keyword{internal}
