% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{map_data}
\alias{map_data}
\title{Reads a dataframe and maps it columns}
\usage{
map_data(data, data_origin = "datalist", custom_map = NULL)
}
\arguments{
\item{data}{data.frame to be transformed}

\item{data_origin}{a string indicating the origin of the data, can be one of
\code{datalist}, \code{data_export}, or \code{custom}}

\item{custom_map}{TODO: Explain column_mapping
TODO: How is the column mapping to be defined? As a list or rather as
parameters?}
}
\value{
a data.frame with the column names as used in the code
}
\description{
\code{map_data} reads a dataframe and maps it columns to match the ones
used in the salary analysis.
}
\details{
TODO: A longer description of what is being done exactly.
}
\keyword{internal}
